+++ admin/settings/meta-box.php
@@ -12,35 +12,40 @@
  * @link https://codex.wordpress.org/Plugin_API/Action_Reference/add_meta_boxes
  */
 
-// get plugin settings
-$options = get_option('lmt_plugin_global_settings');
+add_action( 'admin_init', 'lmt_add_meta_boxes_init' );
 
-function lmt_add_meta_boxes( $post ) {
+function lmt_add_meta_boxes_init( $post ) {
+    // get plugin settings
+    $options = get_option('lmt_plugin_global_settings');
 
-    // If user can't publish posts, then get out
-    if ( ! current_user_can( 'publish_posts' ) ) return;
+    add_action( 'save_post', 'lmt_save_meta_boxes_data', 10, 2 );
+
+    if( isset($options['lmt_enable_last_modified_cb']) && ($options['lmt_enable_last_modified_cb'] == 1) ) {
+        if( isset( $options['lmt_show_last_modified_time_date_post'] ) && $options['lmt_show_last_modified_time_date_post'] != 'manual' ) {
+            
+            add_action( 'add_meta_boxes_post', 'lmt_add_meta_boxes' );
     
-    add_meta_box( 'lmt_meta_box', __( 'WP Last Modified Info', 'wp-last-modified-info' ), 'lmt_meta_box_callback', '', 'side', 'default' );
-}
-
-if( isset($options['lmt_enable_last_modified_cb']) && ($options['lmt_enable_last_modified_cb'] == 1) ) {
-    
-    if( isset( $options['lmt_show_last_modified_time_date_post'] ) && $options['lmt_show_last_modified_time_date_post'] != 'manual' ) {
-        add_action( 'add_meta_boxes_post', 'lmt_add_meta_boxes' );
-
-        if( isset($options['lmt_custom_post_types_list']) ) {
-            $post_types = $options['lmt_custom_post_types_list'];
-            foreach( $post_types as $post_type ) {
-                add_action( "add_meta_boxes_{$post_type}", "lmt_add_meta_boxes" );
+            if( isset($options['lmt_custom_post_types_list']) ) {
+                $post_types = $options['lmt_custom_post_types_list'];
+                foreach( $post_types as $post_type ) {
+                    add_action( "add_meta_boxes_{$post_type}", "lmt_add_meta_boxes" );
+                }
             }
+        }
+    }
+    
+    if( isset($options['lmt_enable_last_modified_page_cb']) && ($options['lmt_enable_last_modified_page_cb'] == 1) ) {
+        if( isset( $options['lmt_show_last_modified_time_date_page'] ) && $options['lmt_show_last_modified_time_date_page'] != 'manual' ) {
+            add_action( 'add_meta_boxes_page', 'lmt_add_meta_boxes' );
         }
     }
 }
 
-if( isset($options['lmt_enable_last_modified_page_cb']) && ($options['lmt_enable_last_modified_page_cb'] == 1) ) {
-    if( isset( $options['lmt_show_last_modified_time_date_page'] ) && $options['lmt_show_last_modified_time_date_page'] != 'manual' ) {
-        add_action( 'add_meta_boxes_page', 'lmt_add_meta_boxes' );
-    }
+function lmt_add_meta_boxes( $post ) {
+    // If user can't publish posts, then get out
+    if ( ! current_user_can( 'publish_posts' ) ) return;
+    
+    add_meta_box( 'lmt_meta_box', __( 'WP Last Modified Info', 'wp-last-modified-info' ), 'lmt_meta_box_callback', '', 'side', 'default' );
 }
 
 /**
@@ -87,7 +92,6 @@ function lmt_meta_box_callback( $post ) {
  * @param int $post_id The post ID.
  */
 function lmt_save_meta_boxes_data( $post_id ) {
-    
     // verify taxonomies meta box nonce
 	if ( ! isset( $_POST['lmt_meta_box_nonce'] ) || ! wp_verify_nonce( $_POST['lmt_meta_box_nonce'], 'lmt_meta_box_build_nonce' ) ) {
 		return;
@@ -107,9 +111,6 @@ function lmt_save_meta_boxes_data( $post_id ) {
     } else {
         delete_post_meta( $post_id, '_lmt_disable' );
     }
-    
 }
-
-add_action( 'save_post', 'lmt_save_meta_boxes_data', 10, 2 );
 
 ?>