+++ includes/frontend/post-options.php
@@ -10,9 +10,7 @@
 add_action( 'init', 'lmt_plugin_init_post' );
 
 function lmt_plugin_init_post() {
-
-    $priority = 10;
-    $priority = apply_filters( 'wplmi_display_priority_post', $priority );
+    $priority = apply_filters( 'wplmi_display_priority_post', 10 );
     
     add_filter( 'the_content', 'lmt_print_last_modified_info_post', $priority );
 }
@@ -63,12 +61,11 @@ function lmt_print_last_modified_info_post( $content ) {
 
     if( isset($options['lmt_custom_post_types_list']) ) {
         $post_types = $options['lmt_custom_post_types_list'];
-        //foreach( $post_types as $post_type ) {
-            if ( isset( $fullcontent ) && is_singular( $post_types ) && !get_post_meta( get_the_ID(), '_lmt_disable', true ) == 'yes' ) { 
-                return $fullcontent;
-            }
-        //}
+        if ( isset( $fullcontent ) && is_singular( $post_types ) && !get_post_meta( get_the_ID(), '_lmt_disable', true ) == 'yes' ) { 
+            return $fullcontent;
+        }
     }
+
     return $content;
 }