+++ includes/theme-support.php
@@ -7,12 +7,11 @@
  * @license   http://www.gnu.org/licenses/gpl.html
  */
 
+$options = get_option('lmt_plugin_global_settings');
+
 /**
  * GeneratePress Theme Support
  */
-
-$options = get_option('lmt_plugin_global_settings');
-
 if( isset($options['lmt_tt_generatepress_theme_mod']) && ($options['lmt_tt_generatepress_theme_mod'] == 'replace') ) {
     add_filter( 'generate_post_date_output', 'lmt_generatepress_theme_replace_meta' );
     add_filter( 'generate_post_author_output', 'lmt_generatepress_theme_replace_author_meta' );
@@ -47,7 +46,6 @@ function lmt_remove_hentry_support_generatepress_ob_callback( $buffer ) {
  * @param String $output Markup for the last modified date.
  * @return void
  */
-
 if( isset($options['lmt_tt_astra_theme_mod']) && ($options['lmt_tt_astra_theme_mod'] == 'replace') ) {
 	add_filter( 'astra_post_date', 'lmt_astra_modified_post_date' );
 }
@@ -56,6 +54,7 @@ function lmt_astra_modified_post_date( $output ) {
 	$output = '';
 	$output .= '<span class="posted-on">';
 	$output .= '<span class="post-updated"> ' . get_the_last_modified_info() . '</span>';
-	$output .= '</span>';
+    $output .= '</span>';
+    
 	return $output;
 }